requirements.txt
openai==1.68.2
python-dotenv==1.0.1
boto3==1.37.18
matplotlib==3.10.1
pandas==2.2.3
numpy==2.2.4
fastapi==0.115.12
uvicorn==0.32.1
tabulate==0.9.0